refactor(server): extract routes.rs test module into routes/tests.rs (#1200)
TD-033 (partial): the route handlers were already split into 8 routes/
submodules, but routes.rs still carried a ~1,626-LOC #[cfg(test)] mod tests.
Move it to routes/tests.rs (declared #[cfg(test)] pub mod tests;), leaving
routes.rs a 73-line module-decl + re-export facade. The module stays pub so
editor_routes tests keep consuming crate::routes::tests::test_app_state.
291 tests pass unchanged.
